dbt/requirements.txt
boto3
pytest
dbt-duckdb
duckdb
pyspark
delta-spark
